Printer Config and Auto-Level Routine
The autolevel routine has been modified to allow for a fixed and probe z
switch.  This removes the need to have the z offset from probe to
extruder at the accuracy necessary for optimal first layer height.

To avoid servo inaccuracy, the probe does not extend or retract between
measurement points when running G29.